+++ CHANGES.md
@@ -14,6 +14,7 @@ Changes in 8.1
 - SQL BILLING_FEES table: Add FILE_ATTACHED column in Update.fnc.php & rosariosis.sql
 - Add File Attached to Fees in StudentFees.php & Student_Billing/functions.inc.php
 - Fix User Widgets for Advanced Search / Print Class Lists exports in GetStaffList.fnc.php, GetStuList.fnc.php & Search.inc.php
+- Add Export fields list (form) & Export fields list + extra SQL (student list) action hooks in Export.php & Actions.php
 
 Changes in 8.0.4
 ----------------